Share Mermaid tooling across web bundles
